various types

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long int m;
 	long long int o;
 	float f;
-		printf("Size of a char: %lu byte(s)\n", sizeof(a));
-		printf("Size of a int: %lu byte(s)\n", sizeof(n));
-		printf("Size of a long int: %lu byte(s)\n", sizeof(m));
-		printf("Size of a long long int: %lu byte(s)\n", sizeof(o));
-		printf("Size of a float: %lu byte(s)\n", sizeof(f));
+		printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+		printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(n));
+		printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(m));
+		printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(o));
+		printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
